CON-359: Split up finality package into `votingmatrix` and `singlesweep`.
