+++ casper/src/main/scala/io/casperlabs/casper/finality/CommitteeWithConsensusValue.scala
@@ -1,0 +1,9 @@
+package io.casperlabs.casper.finality
+
+import io.casperlabs.casper.Estimator.{BlockHash, Validator}
+
+case class CommitteeWithConsensusValue(
+    validator: Set[Validator],
+    quorum: Long,
+    consensusValue: BlockHash
+)
