+++ casper/src/main/scala/io/casperlabs/casper/finality/singlesweep/Committee.scala
@@ -1,0 +1,5 @@
+package io.casperlabs.casper.finality.singlesweep
+
+import io.casperlabs.casper.Estimator.Validator
+
+case class Committee(validators: Set[Validator], quorum: Long)
