+++ casper/src/test/scala/io/casperlabs/casper/helper/BlockGenerator.scala
@@ -11,15 +11,18 @@ import io.casperlabs.blockstorage.{
   IndexedDagStorage
 }
 import io.casperlabs.casper.Estimator.{BlockHash, Validator}
+import io.casperlabs.casper.consensus.Block.ProcessedDeploy
 import io.casperlabs.casper.consensus._
-import Block.ProcessedDeploy
-import io.casperlabs.casper.util.ProtoUtil
-import io.casperlabs.casper.util.execengine.DeploysCheckpoint
-import io.casperlabs.casper.util.execengine.ExecEngineUtil
-import io.casperlabs.casper.util.execengine.ExecEngineUtil.{computeDeploysCheckpoint, StateHash}
 import io.casperlabs.casper.consensus.state.ProtocolVersion
-import io.casperlabs.casper.finality.FinalityDetector.CommitteeWithConsensusValue
-import io.casperlabs.casper.finality.{FinalityDetectorVotingMatrix, VotingMatrix, VotingMatrixImpl}
+import io.casperlabs.casper.finality.CommitteeWithConsensusValue
+import io.casperlabs.casper.finality.votingmatrix.{
+  FinalityDetectorVotingMatrix,
+  VotingMatrix,
+  VotingMatrixImpl
+}
+import io.casperlabs.casper.util.ProtoUtil
+import io.casperlabs.casper.util.execengine.{DeploysCheckpoint, ExecEngineUtil}
+import io.casperlabs.casper.util.execengine.ExecEngineUtil.{computeDeploysCheckpoint, StateHash}
 import io.casperlabs.p2p.EffectsTestInstances.LogicalTime
 import io.casperlabs.shared.{Log, Time}
 import io.casperlabs.smartcontracts.ExecutionEngineService
