+++ casper/src/test/scala/io/casperlabs/casper/helper/GossipServiceCasperTestNode.scala
@@ -13,8 +13,11 @@ import eu.timepit.refined.auto._
 import io.casperlabs.blockstorage._
 import io.casperlabs.casper
 import io.casperlabs.casper.deploybuffer.{DeployBuffer, MockDeployBuffer}
-import io.casperlabs.casper.finality.{FinalityDetector, FinalityDetectorBySingleSweepImpl}
-import io.casperlabs.casper.validation.{Validation, ValidationImpl}
+import io.casperlabs.casper.finality.singlesweep.{
+  FinalityDetector,
+  FinalityDetectorBySingleSweepImpl
+}
+import io.casperlabs.casper.validation.Validation
 import io.casperlabs.casper.{consensus, _}
 import io.casperlabs.comm.CommError.ErrorHandler
 import io.casperlabs.comm.discovery.{Node, NodeDiscovery, NodeIdentifier}
