+++ node/src/main/scala/io/casperlabs/node/api/GrpcCasperService.scala
@@ -10,7 +10,7 @@ import io.casperlabs.casper.api.BlockAPI
 import io.casperlabs.casper.consensus.info._
 import io.casperlabs.casper.consensus.{state, Block}
 import io.casperlabs.casper.deploybuffer.DeployBuffer
-import io.casperlabs.casper.finality.FinalityDetector
+import io.casperlabs.casper.finality.singlesweep.FinalityDetector
 import io.casperlabs.casper.validation.Validation
 import io.casperlabs.catscontrib.MonadThrowable
 import io.casperlabs.comm.ServiceError.InvalidArgument
