+++ node/src/main/scala/io/casperlabs/node/api/Servers.scala
@@ -8,7 +8,7 @@ import io.casperlabs.blockstorage.BlockStorage
 import io.casperlabs.casper.MultiParentCasperRef.MultiParentCasperRef
 import io.casperlabs.casper.deploybuffer.DeployBuffer
 import io.casperlabs.casper.consensus.Block
-import io.casperlabs.casper.finality.FinalityDetector
+import io.casperlabs.casper.finality.singlesweep.FinalityDetector
 import io.casperlabs.casper.protocol.CasperMessageGrpcMonix
 import io.casperlabs.casper.validation.Validation
 import io.casperlabs.comm.discovery.{NodeDiscovery, NodeIdentifier}
